requirements.txt
langgraph>=0.4.5
langchain>=0.3.8
langchain-openai>=0.2.9
langchain-community>=0.3.8
chromadb>=1.0.9
pydantic>=2.9.2
tavily-python>=0.7.2
exa-py>=1.13.1
python-dotenv>=1.0.0
openai>=1.54.0
tiktoken>=0.7.0
numpy>=1.22.4
tqdm>=4.65.0 